drm/lima: Fix lima cache creation
Due to the kernel merge of the hardened usercopy patch set
we have to whitelist the memory area which is used for
copy_from_user actions. Otherwise lima fails with an abort
and segmentation fault.

This is necessary on systems with CONFIG_HARDENED_USERCOPY enabled.
It was tested on a Cubieboard 1, Allwinner A10.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>

%% original patch: 0041-drm-lima-Fix-lima-cache-creation.patch